+++ src/URLResolver.js
@@ -1,12 +1,15 @@
 import http from 'http';
 import https from 'https';
 
-const fetchUrl = url => {
+const fetchUrl = (url, headers = {}) => {
 	return new Promise((resolve, reject) => {
 		const parsedUrl = new URL(url);
 		const h = (parsedUrl.protocol === 'https:') ? https : http;
+		let options = {
+			headers: headers
+		};
 
-		h.get(url, res => {
+		h.get(url, options, res => {
 			if (res.statusCode >= 300 && res.statusCode < 400 && res.headers.location) { // redirect url
 				fetchUrl(res.headers.location).then(buffer => {
 					resolve(buffer);
@@ -34,11 +37,11 @@ class URLResolver {
 		this.resolving = {};
 	}
 
-	resolve(url) {
+	resolve(url, headers = {}) {
 		if (!this.resolving[url]) {
 			this.resolving[url] = new Promise((resolve, reject) => {
 				if (url.toLowerCase().indexOf('https://') === 0 || url.toLowerCase().indexOf('http://') === 0) {
-					fetchUrl(url).then(buffer => {
+					fetchUrl(url, headers).then(buffer => {
 						this.fs.writeFileSync(url, buffer);
 						resolve();
 					}, result => {