+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- attachments embedding
+- Attachments embedding
+- Support passing headers to request for loading font files and images via URL adresses
 
 ## 0.3.0-beta.1 - 2022-01-01